is{Wide,Combining}Char: fast-bail for low-CP text
The lowest wide character is U+1100, and the lowest combining character
is U+0300.  That leaves a lot of code points, down to U+0000, that can't
be wide or combining.  Test for those manually so we don't have to
search the tables for characters that can't possibly be in the tables.